+++ app/src/StagesView.tsx
